Moved reader class to own file
